requirements.txt
asgiref==3.2.7
Django==3.0.5
djangorestframework==3.11.0
gunicorn==20.0.4
numpy==1.18.2
pandas==1.0.3
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
Markdown==2.6.9
Pillow==2.3.0


